requirements.txt
numpy==2.3.1
pyzmq==27.0.0
